requirements.txt
bs4
cloudscraper
pycryptodome

